utils format changed
