+++ utils/formats/getFormatRoute.utility.ts
@@ -1,0 +1,1 @@
+export const getFormatRoute = (pathname:string):string => pathname.split('/').pop() as string
