+++ utils/validators/crearLicitacion.ts
@@ -1,7 +1,7 @@
-import { convertToDate } from '..'
 import { REGEX_FORM } from '../../consts/regex'
 import { FormCrearLicitacionUser } from '../../types/form'
 import { ErrorsForm, ValidatorForm } from '../../types/validators'
+import { convertToDate } from '../formats'
 
 const validatorCrearLicitacion:ValidatorForm<FormCrearLicitacionUser, Omit<FormCrearLicitacionUser, 'tipoLicitacion'| 'meses'|'description'|'requisitos'>> = (form) => {
   const errors:ErrorsForm<Omit<FormCrearLicitacionUser, 'tipoLicitacion'| 'meses'|'description'|'requisitos'>> = {
